refactor(broker): split chain_monitor_v2.rs into directory layout (C38 of PR-3)
Per the canonical plan, chain_monitor.rs becomes legacy.rs inside
chain_monitor_v2/ and block_history.rs is absorbed too. The shared
trait/types live in a new types.rs so both V2 and legacy can reach them.

chain_monitor_v2.rs (982) + chain_monitor.rs (255) + block_history.rs (134)
become:

- chain_monitor_v2/mod.rs (~40 lines): module decls + pub(crate) use
  service::ChainMonitorV2 + pub(crate) use legacy::ChainMonitorService
  + pub(crate) use types::{ChainHead, ChainMonitorObj}. The flat surface
  means external callers use crate::chain_monitor_v2::ChainMonitorService
  rather than ::legacy::ChainMonitorService.
- chain_monitor_v2/error.rs (~49 lines): ChainMonitorV2Err.
- chain_monitor_v2/types.rs (~40 lines): shared ChainHead struct,
  ChainMonitorApi trait, ChainMonitorObj alias — implemented by both
  V2 and legacy.
- chain_monitor_v2/block_history.rs (134 lines): absorbed verbatim from
  top-level; only V2 used it.
- chain_monitor_v2/legacy.rs (~243 lines): ChainMonitorErr +
  ChainMonitorService<P> + its impls + the smoke test. Slated for
  removal once V2 has fully replaced it; lives here to be deleted with
  one rm.
- chain_monitor_v2/service.rs (~952 lines): ChainMonitorV2<P, ANP> +
  BrokerService impl + the V2 tests. Test fixture include_str! paths
  bumped from ../testdata/ to ../../testdata/ for the new depth.

External crate::chain_monitor::* paths all migrate to
crate::chain_monitor_v2::*: broker.rs, aggregator/service.rs,
market_monitor/service.rs, order_locker/service.rs, order_pricer/service.rs.

ChainMonitorApi trait is not re-exported from mod.rs — external code
only references the trait through Arc<dyn ChainMonitorApi> via
ChainMonitorObj.

lib.rs loses the mod chain_monitor and mod block_history declarations.
