+++ crates/broker/src/lib.rs
@@ -21,9 +21,7 @@ pub(crate) use storage::ConfigurableDownloader;
 pub(crate) mod aggregator;
 pub use aggregator::{AggregationState, Batch, BatchStatus};
 pub mod args;
-pub(crate) mod block_history;
 mod broker;
-pub(crate) mod chain_monitor;
 pub(crate) mod chain_monitor_v2;
 pub(crate) mod channels;
 mod db;
